Work in progress

+++ src/margin.cpp
@@ -2,34 +2,34 @@
 
 using namespace geometry;
 
-int32_t margin::top() const noexcept {
+float margin::top() const noexcept {
   return _top;
 }
 
-void margin::set_top(int32_t value) noexcept {
+void margin::set_top(float value) noexcept {
   _top = value;
 }
 
-int32_t margin::left() const noexcept {
+float margin::left() const noexcept {
   return _left;
 }
 
-void margin::set_left(int32_t value) noexcept {
+void margin::set_left(float value) noexcept {
   _left = value;
 }
 
-int32_t margin::bottom() const noexcept {
+float margin::bottom() const noexcept {
   return _bottom;
 }
 
-void margin::set_bottom(int32_t value) noexcept {
+void margin::set_bottom(float value) noexcept {
   _bottom = value;
 }
 
-int32_t margin::right() const noexcept {
+float margin::right() const noexcept {
   return _right;
 }
 
-void margin::set_right(int32_t value) noexcept {
+void margin::set_right(float value) noexcept {
   _right = value;
 }
